Renaming and Rejiggin to nuget properly
