+++ Neo4jClient.Full/Neo4jClient.Full.csproj.DotSettings
