+++ Neo4jClient.Full/Transactions/ClosedTransactionException.cs
