+++ Neo4jClient.Full/Transactions/Neo4jTransactionProxy.cs
