+++ Neo4jClient.Full/Transactions/Neo4jTransactionResourceManager.cs
