+++ Neo4jClient.Full/Transactions/SuppressTransactionProxy.cs
