+++ Neo4jClient.Full/Transactions/TransactionConnectionContext.cs
