+++ Neo4jClient.Full/Transactions/TransactionManager.cs
