+++ Neo4jClient.Full/Transactions/TransactionScopeProxy.cs
