+++ Neo4jClient.Full/Transactions/TransactionSinglePhaseNotification.cs
