--- Neo4jClient.Pcl/Execution/GraphClientBasedExecutionPolicy.cs
@@ -1,12 +1,0 @@
-﻿namespace Neo4jClient.Execution
-{
-    internal abstract partial class GraphClientBasedExecutionPolicy : IExecutionPolicy
-    {
-        public bool InTransaction
-        {
-            get { return false; }
-        }
-
-        public abstract TransactionExecutionPolicy TransactionExecutionPolicy { get; }
-    }
-}
