+++ Neo4jClient.nuspec
@@ -1,0 +1,52 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
+  <metadata>
+    <id>Neo4jClient</id>
+    <version>$version$</version>
+    <authors>Readify, Chris Skardon</authors>
+    <owners>Tatham Oddie, Chris Skardon</owners>
+    <licenseUrl>https://github.com/Readify/Neo4jClient/blob/master/LICENSE</licenseUrl>
+    <projectUrl>https://github.com/Readify/Neo4jClient</projectUrl>
+    <description>A .NET client for neo4j: an open source, transactional graph database. It's pretty awesome.</description>
+    <tags>neo4j nosql cypher</tags>
+    <dependencies>
+      <group targetFramework="dotnet">
+        <dependency id="Microsoft.CSharp" version="4.0.0" />
+        <dependency id="Microsoft.VisualBasic" version="10.0.0" />
+        <dependency id="Newtonsoft.Json" version="8.0.2" />
+        <dependency id="System.Collections" version="4.0.10" />
+        <dependency id="System.Collections.NonGeneric" version="4.0.0" />
+        <dependency id="System.ComponentModel" version="4.0.0" />
+        <dependency id="System.ComponentModel.Annotations" version="4.0.10" />
+        <dependency id="System.ComponentModel.TypeConverter" version="4.0.0" />
+        <dependency id="System.Diagnostics.Debug" version="4.0.10" />
+        <dependency id="System.Dynamic.Runtime" version="4.0.10" />
+        <dependency id="System.Globalization" version="4.0.10" />
+        <dependency id="System.IO" version="4.0.10" />
+        <dependency id="System.Linq" version="4.0.0" />
+        <dependency id="System.Linq.Expressions" version="4.0.10" />
+        <dependency id="System.Net.Http" version="4.0.0" />
+        <dependency id="System.Net.Primitives" version="4.0.10" />
+        <dependency id="System.Reflection" version="4.0.10" />
+        <dependency id="System.Reflection.Extensions" version="4.0.0" />
+        <dependency id="System.Reflection.TypeExtensions" version="4.0.0" />
+        <dependency id="System.Resources.ResourceManager" version="4.0.0" />
+        <dependency id="System.Runtime" version="4.0.20" />
+        <dependency id="System.Runtime.Extensions" version="4.0.10" />
+        <dependency id="System.Runtime.Serialization.Primitives" version="4.0.10" />
+        <dependency id="System.Text.Encoding" version="4.0.10" />
+        <dependency id="System.Text.RegularExpressions" version="4.0.10" />
+        <dependency id="System.Threading" version="4.0.10" />
+        <dependency id="System.Threading.Tasks" version="4.0.10" />
+      </group>
+    </dependencies>
+  </metadata>
+  <files>
+    <file src="Neo4jClient.Full\bin\Release\Neo4jClient.dll" target="lib\net45" />
+    <file src="Neo4jClient.Full\bin\Release\Neo4jClient.pdb" target="lib\net45" />
+    <file src="Neo4jClient.Full\bin\Release\Neo4jClient.xml" target="lib\net45" />
+    <file src="Neo4jClient\bin\Release\Neo4jClient.dll" target="lib\dotnet" />
+    <file src="Neo4jClient\bin\Release\Neo4jClient.pdb" target="lib\dotnet" />
+    <file src="Neo4jClient\bin\Release\Neo4jClient.xml" target="lib\dotnet" />
+  </files>
+</package>
