--- Neo4jClient/Neo4jClient.nuspec
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata>
-    <id>Neo4jClient</id>
-    <version>0.0.0.0</version>
-    <authors>Readify</authors>
-    <owners>Tatham Oddie</owners>
-    <licenseUrl>https://github.com/Readify/Neo4jClient/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/Readify/Neo4jClient</projectUrl>
-    <description>A .NET client for neo4j: an open source, transactional graph database. It's pretty awesome.</description>
-    <tags>neo4j nosql cypher</tags>
-  </metadata>
-  <files>
-    <file src="bin\Release\Neo4jClient.dll" target="lib\net45" />
-    <file src="bin\Release\Neo4jClient.pdb" target="lib\net45" />
-    <file src="bin\Release\Neo4jClient.xml" target="lib\net45" />
-  </files>
-</package>
